Update Chinook Data Analytics
Utilized the Chinook database to investigate and respond to inquiries by constructing various queries. Successfully retrieved relevant data by selecting and joining tables, and filtered by specific criteria when necessary. Analyzed and manipulated data to provide insights and answer questions, such as identifying the sales agent with the highest sales in a specific year. Created queries that included purchased track and artist names with each invoice line ID, as well as all tracks with album name, media type, and genre. Grouped and sorted data to provide a comprehensive view of the information.